organized logic into organized folders/files, dev site now blank
