+++ src/AboutUs/AboutUs.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import '../AboutUs/AboutUs.css'
+
+<div className="about">
+<h1>About Us</h1>
+  <p>Here at Sacred Visionaries, we pride ourselves in holistic fitness! We understand that your journey is unique. <br />
+    Redefining your physical, mental, and spiritual relationships isn't easy. That's what we're here for! To guide those willing to <br />
+    change their lives. From begginer to advanced, weights or calisthenics, and mindset to nutrition, we can help!<br />
+    
+
+  </p>
+</div>
+
+export default About
