+++ src/Testimonials/testimonials.jsx
@@ -1,0 +1,24 @@
+import React from 'react'
+
+<div className="testimonials">
+        <h1>Testimonials</h1>
+        <p>"Working with Martin has definetly given me a different outook on life!<br/>
+        Not only have I seen massive change on my body, my way of thinking changed. Thanks Martin!"<br/>
+        -John Doe</p>
+        <br/>
+        <p>"Rudy has been an inspiration since day one! My relationship with<br/>
+        myself and with God has grown massively and I am forever grateful!"<br/>
+        -John Doe</p>
+        <br/>
+        <p>"Sam is a gamechanger! I feel and look way better than 3 months ago.<br/>
+        She changed the way I view food and the way I view myself!"<br/>
+        </p>
+        <br/>
+        <p>"I never knew calisthenics would be so fun! Yes, it's super challenging,<br/>
+        but my body can do things it wasn't able to years ago! Thanks Joey for being<br/>
+        a great coach!"<br/>
+        -John Doe</p>
+        </div>
+
+
+export default Testimonials
